pyproject.toml
[project]
name = "vaf-alpha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cachetools>=7.0.5",
    "cerebras>=0.0.1",
    "cloverlabs-camoufox",
    "comtypes>=1.4.16",
    "curl-cffi>=0.15.0",
    "httpx[http2]>=0.28.1",
    "icalendar>=7.0.3",
    "ics>=0.7.3",
    "mistralai>=2.1.3",
    "mss>=10.1.0",
    "numpy>=2.4.3",
    "openwakeword>=0.4.0",
    "orjson>=3.11.8",
    "pillow>=12.2.0",
    "pvporcupine>=4.0.2",
    "pvrecorder>=1.2.7",
    "python-dotenv>=1.2.2",
    "pyttsx3>=2.99",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
    "spotipy>=2.26.0",
    "yaspin>=3.4.0",
]
